+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - registry/gc: Add Prometheus counter and histogram for online GC deletions
 - registry/gc: Add Prometheus counter for online GC deleted bytes
 - registry/gc: Add Prometheus counter for online GC review postpones
+- registry/gc: Add Prometheus histogram for sleep durations between online GC runs
 
 ## [v3.3.0-gitlab] - 2021-04-09
 ### Added